+++ src/components/blogs/BlogCard.astro
@@ -3,11 +3,17 @@ import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 import type { PostFrontMatter } from "src/types";
 export type Props = PostFrontMatter;
-const {title, date, slug} = Astro.props;
-dayjs.extend(relativeTime)
+const { title, date, slug } = Astro.props;
+dayjs.extend(relativeTime);
 ---
 
-<div class="border rounded-md border-slate-500/50 p-4 hover:shadow-lg hover:shadow-orange-500 hover:dark:shadow-teal-500 transition-shadow duration-300">
-  <a href={`/blogs/${slug}`} class="text-xl font-bold hover:text-orange-500 hover:dark:text-teal-500 transition">{title}</a>
+<div
+  class="border rounded-md border-slate-500/50 p-4 hover:shadow-lg hover:shadow-orange-500 hover:dark:shadow-teal-500 transition-shadow duration-300"
+>
+  <a
+    href={`/blogs/${slug}`}
+    class="text-xl font-bold hover:text-orange-500 hover:dark:text-teal-500 transition"
+    >{title}
+  </a>
   <p class="py-2">{dayjs(date).fromNow()}</p>
 </div>
