+++ src/components/common/UnderlineLink.astro
@@ -4,7 +4,7 @@ export interface Props {
   title: string;
 }
 
-const {path, title} = Astro.props;
+const { path, title } = Astro.props;
 ---
 
 <a
