+++ src/components/home/About.astro
@@ -1,6 +1,7 @@
 ---
 import Icon from "astro-icon";
-const aStyle = "w-6 h-6 hover:text-orange-500 hover:dark:text-teal-600 transition"
+const aStyle =
+  "w-6 h-6 hover:text-orange-500 hover:dark:text-teal-600 transition";
 ---
 
 <style>
@@ -11,25 +12,35 @@ const aStyle = "w-6 h-6 hover:text-orange-500 hover:dark:text-teal-600 transitio
 
 <div class="flex px-4 py-10 items-center gap-6 justify-between h-screen">
   <div class="flex flex-col space-y-2">
-    <p class="font-bold text-5xl">Hello! I am <span class="text-orange-500 dark:text-teal-500">Viet.</span></p>
-    <p class="py-2">I'm an international student at <span>Simon Fraser University</span>.
-      I enjoy creating simple, beautiful and user-friendly web applications and I'm passionate to learn full-stack web technology.</p>
-      
-      <p>I'm currently looking for a role as a developer.</p>
-      <div class="flex items-center space-x-4 py-2">
-        <button
-          class="p-2 rounded-lg text-white bg-orange-500 dark:bg-teal-600 hover:show-md hover:dark:bg-teal-700 transition inline-flex gap-2 items-center"
-        >
-          <Icon class="w-4 h-4" name="fa-solid:download" /> Resumé
-        </button>
-        <a href="Github">
-          <Icon class={aStyle} name="fa-brands:github" />
-        </a>
-        <a href="Linkedin">
-          <Icon class={aStyle} name="fa-brands:linkedin" />
-        </a>
-      </div>
-    </div>
+    <p class="font-bold text-5xl">
+      Hello! I am <span class="text-orange-500 dark:text-teal-500">Viet.</span>
+    </p>
+    <p class="py-2">
+      I'm an international student at <span>Simon Fraser University</span>. I
+      enjoy creating simple, beautiful and user-friendly web applications and
+      I'm passionate to learn full-stack web technology.
+    </p>
 
-    <img alt="My photo" width="300px" class="hidden md:block rounded-xl shadow-xl shadow-orange-500 dark:shadow-teal-500" src="/assets/profile.jpg" />
+    <p>I'm currently looking for a role as a developer.</p>
+    <div class="flex items-center space-x-4 py-2">
+      <button
+        class="p-2 rounded-lg text-white bg-orange-500 dark:bg-teal-600 hover:show-md hover:dark:bg-teal-700 transition inline-flex gap-2 items-center"
+      >
+        <Icon class="w-4 h-4" name="fa-solid:download" /> Resumé
+      </button>
+      <a href="Github">
+        <Icon class={aStyle} name="fa-brands:github" />
+      </a>
+      <a href="Linkedin">
+        <Icon class={aStyle} name="fa-brands:linkedin" />
+      </a>
+    </div>
   </div>
+
+  <img
+    alt="My photo"
+    width="300px"
+    class="hidden md:block rounded-xl shadow-xl shadow-orange-500 dark:shadow-teal-500"
+    src="/assets/profile.jpg"
+  />
+</div>
