+++ src/components/home/RecentBlogs.astro
@@ -2,17 +2,20 @@
 import type { PostFrontMatter } from "src/types";
 import BlogCard from "../blogs/BlogCard.astro";
 import Link from "../common/UnderlineLink.astro";
-const posts = await Astro.glob<PostFrontMatter>('../../content/posts/*.mdx');
+const posts = await Astro.glob<PostFrontMatter>("../../content/posts/*.mdx");
 ---
 
 <div class="px-4 py-12">
   <h2 class="text-3xl font-bold mb-4">Recent Blogs</h2>
-  <p class="text-black/50 dark:text-white/50">A records of my learning journey into coding.
+  <p class="text-black/50 dark:text-white/50">
+    A records of my learning journey into coding.
     <Link path="/blogs" title="View all blogs" />
   </p>
   <div class="grid grid-cols-1 md:grid-cols-2 gap-4 mt-6">
-    {posts.slice(0, 4).map((post) => {
-      return <BlogCard {...post.frontmatter} />
-    })}
+    {
+      posts.slice(0, 4).map((post) => {
+        return <BlogCard {...post.frontmatter} />;
+      })
+    }
   </div>
 </div>
