+++ src/content/projects/wordle.md
@@ -6,6 +6,6 @@ demo: https://wordle-solid.vercel.app
 important: true
 ---
 
-A clone of popular game [Wordle](https://www.nytimes.com/games/wordle/index) 
+A clone of popular game [Wordle](https://www.nytimes.com/games/wordle/index)
 built with [SolidJS](https://www.solidjs.com/) and [Tailwind CSS](https://tailwindcss.com/).
 Additionally, the game supports auto-save and refreshes daily.
