+++ src/pages/blogs/[slug].astro
@@ -6,42 +6,50 @@ import dayjs from "dayjs";
 import type { PostFrontMatter } from "src/types";
 
 export async function getStaticPaths() {
-  const posts = await Astro.glob<PostFrontMatter>('../../content/posts/**/*.mdx');
-  return posts.map(post => ({
+  const posts = await Astro.glob<PostFrontMatter>(
+    "../../content/posts/**/*.mdx"
+  );
+  return posts.map((post) => ({
     params: {
       slug: post.frontmatter.slug
     },
     props: {
       post
     }
-  }))
+  }));
 }
 
 export type Props = {
   post: MDXInstance<PostFrontMatter>;
-}
+};
 
 const { Content, frontmatter, getHeadings } = Astro.props.post;
 
-const date = dayjs(frontmatter.date).format("D MMMM YYYY")
+const date = dayjs(frontmatter.date).format("D MMMM YYYY");
 ---
 
 <BaseLayout title={frontmatter.title}>
-  <main class="flex flex-1 px-4 max-w-screen-lg mx-auto gap-4 items-start py-24">
+  <main
+    class="flex flex-1 px-4 max-w-screen-lg mx-auto gap-4 items-start py-24"
+  >
     <article class="flex flex-col flex-1 mx-auto gap-4">
       <h1 class="text-4xl font-bold">{frontmatter.title}</h1>
       <h5 class="text-black/50 dark:text-white/50">{date}</h5>
       <div class="flex gap-4">
-        <div class="flex-1 prose prose-sm max-w-none prose-orange dark:prose-invert dark:prose-teal prose-headings:scroll-mt-24">
+        <div
+          class="flex-1 prose prose-sm max-w-none prose-orange dark:prose-invert dark:prose-teal prose-headings:scroll-mt-24"
+        >
           <Content />
         </div>
       </div>
     </article>
     <aside
-      class="w-60 hidden md:flex flex-col gap-3 sticky top-24 bg-slate-500/10 p-4 rounded-md items-start">
-    <TableOfContent 
-      client:load 
-      headings={getHeadings().filter(heading => heading.depth <= 3)} />
+      class="w-60 hidden md:flex flex-col gap-3 sticky top-24 bg-slate-500/10 p-4 rounded-md items-start"
+    >
+      <TableOfContent
+        client:load
+        headings={getHeadings().filter((heading) => heading.depth <= 3)}
+      />
     </aside>
   </main>
 </BaseLayout>
